feat: Add forcing option to Dataset initialization
The code changes add a new `forcing` option to the `Dataset` class initialization, allowing the user to enable forcing in the dataset. This option is set to `False` by default. Additionally, a new `n_forcing` parameter is added to specify the number of forcing points. The commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).